mac: Minor bug fix: 2.5 cannot be stored in an int.
Found by clang's newly enabled -Wliteral-conversion:

../../chrome/browser/ui/cocoa/download/download_item_cell.mm:75:40: error: implicit conversion turns literal floating-point number into integer: 'double' to 'const int' [-Werror,-Wliteral-conversion]
const int kCompleteAnimationDuration = 2.5;
          ~~~~~~~~~~~~~~~~~~~~~~~~~~   ^~~
../../chrome/browser/ui/cocoa/download/download_item_cell.mm:78:43: error: implicit conversion turns literal floating-point number into integer: 'double' to 'const int' [-Werror,-Wliteral-conversion]
const int kInterruptedAnimationDuration = 2.5;
          ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~   ^~~
2 errors generated.

BUG=none
TEST=download animation runs 0.5s longer.


Review URL: http://codereview.chromium.org/10082026

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@132324 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 6542a57b3ab7317f5bbf06990ce98bfb259bc7af

+++ chrome/browser/ui/cocoa/download/download_item_cell.mm
@@ -72,10 +72,10 @@ NSTimeInterval kShowStatusDuration = 0.3;
 NSTimeInterval kHideStatusDuration = 0.3;
 
 // Duration of the 'download complete' animation, in seconds.
-const int kCompleteAnimationDuration = 2.5;
+const CGFloat kCompleteAnimationDuration = 2.5;
 
 // Duration of the 'download interrupted' animation, in seconds.
-const int kInterruptedAnimationDuration = 2.5;
+const CGFloat kInterruptedAnimationDuration = 2.5;
 
 using content::DownloadItem;
